pyproject.toml
[project]
name = "voiceflow-docs-mcp"
version = "0.1.0"
description = "MCP server for Voiceflow documentation search"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "fastmcp>=2.13.0.2",
    "httpx>=0.28.1",
    "markdownify>=1.2.0",
    "playwright>=1.56.0",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.3",
]

[project.scripts]
voiceflow-docs-mcp = "voiceflow_docs_mcp.server:main"
